ARM: pxa: Add mapping from GPIO no. to pinctrl pin no.
Currently there is no mapping from GPIO no. to pinctrl pin no. This
causes GPIO API's to fail, since they internally invoke the pinctrl
API, which needs the mapping. In other drivers this mapping is done as
part of device tree. But since we are doing this without the device
tree, we are doing the mapping ourselves.

This is essentially a HACK, and I am not sure of the right way to do
it within the current setup.

Signed-off-by: Vijay Kumar B. <vijaykumar@zilogic.com>